+++ internal/packages/pacman.go
@@ -3,6 +3,7 @@ package packages
 import (
 	"bufio"
 	"bytes"
+	"os"
 	"os/exec"
 	"strings"
 )
@@ -85,8 +86,9 @@ func (p *PacmanManager) Install(pkgs []string) error {
 	}
 	args := append([]string{"-S", "--noconfirm"}, pkgs...)
 	cmd := exec.Command("sudo", append([]string{"pacman"}, args...)...)
-	cmd.Stdout = nil
-	cmd.Stderr = nil
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	cmd.Stdin = os.Stdin
 	return cmd.Run()
 }
 
@@ -96,7 +98,8 @@ func (p *PacmanManager) Uninstall(pkgs []string) error {
 	}
 	args := append([]string{"-R", "--noconfirm"}, pkgs...)
 	cmd := exec.Command("sudo", append([]string{"pacman"}, args...)...)
-	cmd.Stdout = nil
-	cmd.Stderr = nil
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	cmd.Stdin = os.Stdin
 	return cmd.Run()
 }